% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_matches.R
\name{cbind_matches}
\alias{cbind_matches}
\title{Combine data frames column-wise by matching rows}
\usage{
cbind_matches(df1, df2, ind_col)
}
\arguments{
\item{df1}{data frame}

\item{df2}{data frame}

\item{ind_col}{character value giving the index column name}
}
\description{
Finds all rows in two data frames that have matching
entries for an index column and returns a data frame with only
these rows combined, ordered by the index column entries (as
\code{order()} would order them).
}
\examples{
df1 <- data.frame(z = c(3, 2, 8, 6), y = c("foo", "bar", "spam", "mayo"))
df2 <- data.frame(x = c(310, 580, 43, 7), y = c("top", "foo", "hurr", "bar"))
cbind_matches(df1, df2, "y")

}
\seealso{
\code{\link[base]{order}}
}
