% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_usern.R
\name{clean_usern}
\alias{clean_usern}
\title{Clean up RATSS user names}
\usage{
clean_usern(vec)
}
\arguments{
\item{vec}{character vector}
}
\description{
Takes a character vector of  entries that
have a user name in them in the format "<digit/digits>_<single digit>"
somewhere in them, extracts only the user names and returns a vector
with those.
}
\seealso{
\code{\link[stringr]{str_extract}}
}
