% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_scoring.R
\name{consistency_scoring}
\alias{consistency_scoring}
\title{Calculate consistency scores}
\usage{
consistency_scoring(
  part_df,
  id_col_name,
  timestamp_col_name,
  plotdir = NULL,
  method = "euclidean",
  fmt = "Luv",
  nameby = NULL,
  swedish_chars = FALSE,
  swedish_weekdays = FALSE,
  participant_ids = NULL,
  category_lines = FALSE
)
}
\arguments{
\item{part_df}{participant consistency test raw data frame}

\item{plotdir}{(optional) directory path for directory to put plots in}

\item{method}{method to use in color space distance calculations
("euclidean" for Euclidean distances, "manhattan" for
Manhattan/city block distances - "euclidean" is standard)}

\item{fmt}{specifies color format/space to be used ("sRGB", "Luv")}

\item{nameby}{(optional) name, in quotes, of column to name output plots by, e. g. "USERNAME"}

\item{swedish_chars}{specify if swedish characters are in the data frame or not}

\item{swedish_weekdays}{specify if swedish weekdays are in the data frame or not}

\item{participant_ids}{a vector for naming output plots. needs to be in the same order as the part_df participants.}

\item{category_lines}{if TRUE and plotdir is set, drawn plots will include lines
representing the mean score per category}
}
\description{
Calculates consistency scores using data from a consistency
test for synesthesia. Can also create plots for each data set (using the
synratss::plot_syn_cons function).
}
\seealso{
\code{\link[synratss]{prop_col}} \code{\link[synratss]{dist_sum}}
\code{\link[synratss]{plot_syn_cons}}
}
