% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_scoring_swe.R
\name{consistency_scoring_swe}
\alias{consistency_scoring_swe}
\title{Calculate consistency scores (Swe ver.)}
\usage{
consistency_scoring_swe(
  part_df,
  plotdir = NULL,
  method = "euclidean",
  fmt = "Luv",
  nameby = NULL
)
}
\arguments{
\item{part_df}{participant consistency test raw data frame}

\item{plotdir}{(optional) directory path for directory to put plots in}

\item{method}{method to use in color space distance calculations
("euclidean" for Euclidean distances, "manhattan" for
Manhattan/city block distances - "euclidean" is standard)}

\item{fmt}{specifies color format/space to be used ("sRGB", "Luv")}

\item{nameby}{(optional) name, in quotes, of column to name
output plots by, e. g. "USERNAME"}
}
\description{
This function is identical to \code{consistency_scoring},
except the plots it produces are in Swedish and adjusted for participants.
}
\seealso{
\code{\link[synratss]{consistency_scoring}}
}
