% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cspace.R
\name{hex_to_cspace}
\alias{hex_to_cspace}
\title{Convert hex code to color space}
\usage{
hex_to_cspace(hexcobj, fmt = "sRGB")
}
\arguments{
\item{hexcobj}{a data frame (single row) or character vector with hexadecimal color values}

\item{fmt}{specifies color axes/output format ("sRGB", "Luv", "Lab").
note that output sRGB values are on the scale 0-1}
}
\description{
Converts hexadecimal color specifications to different
color-space representations, where columns represent different responses/colors,
and each row represents one color axis (e.g. R-G-B)
}
\seealso{
\code{\link[grDevices]{col2rgb}}
\code{\link[grDevices]{convertColor}}
}
