% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_color.R
\name{is_color}
\alias{is_color}
\title{Check if hex color code falls within specified color range}
\usage{
is_color(hex, col = NULL, r = NULL, g = NULL, b = NULL)
}
\arguments{
\item{hex}{hexadecimal color value to check (e. g. "#DA80F2")}

\item{col}{color specification (possible choices are: "blue", "red",
"green", "white", "black", "hazy")}

\item{r}{a two-value vector}

\item{g}{a two-value vector}

\item{b}{a two-value vector}
}
\description{
Checks if a hexadecimal color value codes for a color
within a specified color range. Color range is specified using EITHER col OR
r/g/b arguments. Possible col specifications are: "blue", "red", "green",
"white", "black" or "hazy". r/g/b each take a
two-value numeric vector providing
specifications of minimum/maximum red/green/blue values within the
range 0-255 (e.g. c(20, 230) ).
}
\seealso{
\code{\link[grDevices]{col2rgb}}
}
