% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_syn_cons_swe.R
\name{plot_syn_cons_swe}
\alias{plot_syn_cons_swe}
\title{Plot consistency scores and responses by synesthetic inducer (Swe ver.)}
\usage{
plot_syn_cons_swe(ddf, hexcolors, savepath = NULL, tovar = FALSE)
}
\arguments{
\item{ddf}{vector/single-row array of mean difference between response colors.}

\item{hexcolors}{data frame with hexadecimal color codes for responses
for all graphemes}

\item{savepath}{(optional) file path for saving produced plot}

\item{tovar}{tovar: if TRUE, resulting plot is returned so that you
can save it to a variable. defaults to FALSE}
}
\description{
This function is identical to \code{plot_syn_cons},
except the plots it produces are in Swedish and adjusted for participants.
}
\seealso{
\code{\link[synratss]{plot_syn_cons}}
}
