% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnr_to_bdate.R
\name{pnr_to_bdate}
\alias{pnr_to_bdate}
\title{Convert a personal number to birth date}
\usage{
pnr_to_bdate(pnr)
}
\arguments{
\item{pnr}{a string/character value with a single personal number}
}
\description{
Converts Swedish personal numbers (in the format "YYYYMMDD-XXXX",
or "YYMMDD-XXXX") to birth dates. If the personal number is in the format
"YYMMDD-XXXX", the function assumes that all YY values above 25 indicate
someone born in the 1900's.
}
\examples{
x <- "19400830-1111"
pnr_bdate(x)
y <- "090830-1111"
pnr_bdate(y)

}
\seealso{
\code{\link[base]{sub}}
}
