% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_col.R
\name{prop_col}
\alias{prop_col}
\title{Calculate proportion of color codes within specified color range}
\usage{
prop_col(hexdf, col = "black", byrow = FALSE)
}
\arguments{
\item{hexdf}{data frame with hexadecimal color values}

\item{col}{1-value character vector, specifying color range}

\item{byrow}{1-value character vector, specifying if function should calculate
the proportion of rows (TRUE) where all values fall within color range,
or the proportion of elements (FALSE - default).}
}
\description{
Calculates the proportion of color values in a data frame that
fall within a color range, specified using the col argument. Possible choices
are "black", "white", "blue" and "hazy", with "black" as default. Returns NA
if all values in the data frame are NA.
}
\seealso{
\code{\link[synratss]{is_color}}
}
