% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratid_to_usern.R
\name{ratid_to_usern}
\alias{ratid_to_usern}
\title{Convert a RATSS study ID to username format}
\usage{
ratid_to_usern(ratid)
}
\arguments{
\item{ratid}{a string/character vector with a single RATSS code}
}
\description{
Converts a RATSS code/study ID (in the format e.g.
"11 0104 01 01 14" or without any spaces "110104010114") to a
<family>_<twin_no> format (e.g. "104_1"). For parents, this leads to them
having usernames in the format "<family no>_5" or "<family no>_6". Note that
this parent coding might not be what you want or
what has been used previously.
}
\examples{
x <- "11 0104 01 01 14"
ratid_to_usern(x)
y <- "110104010114"
ratid_to_usern(y)

}
\seealso{
\code{\link[base]{gsub}} \code{\link[base]{substr}}
}
