% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synratss.R
\docType{package}
\name{synratss-package}
\alias{synratss}
\alias{synratss-package}
\title{synratss: Handles data from synesthesia consistency test}
\description{
The main function is consistency_scoring, which uses the other
    functions in the package. It calculates consistency scores for data from
    an online consistency test (see e.g. Rothen, Seth, Witzel, & Ward, 2013)
    and produces plots for each data set/participant. This package was
    developed for use at the Center of Neurodevelopmental Disorders at
    Karolinska Institutet (KIND), under the supervision of Janina Neufeld. Its
    applicability is highly narrow and specific.
}
\details{
This package is used at KIND for synesthesia-related research.
}
\keyword{KIND}
\keyword{RATSS}
\keyword{internal}
\keyword{synesthesia}
\keyword{twins}
