% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synratss.R
\docType{package}
\name{synratss}
\alias{synratss}
\title{synratss}
\description{
This package was used at KIND for analyzing and visualizing synesthesia
consistency test data and has been deprecated.
You are now recommended to instead use the synr package, which can be
found at www.github.com/datalowe/synr.
}
\details{
The main function is \code{consistency_scoring()}, which uses the other
functions in the package. It calculates consistency scores for data from
an online consistency test (see e.g. Rothen, Seth, Witzel, & Ward, 2013)
and produces plots for each data set/participant. You can also use the
\code{plot_syn_cons()} function on its own (it is otherwise called by the
\code{consistency_scoring()} function).

This package was developed for use at the Center of
Neurodevelopmental Disorders at Karolinska Institutet (KIND). Its
applicability is highly narrow and specific.
}
