% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_scoring.R
\name{consistency_scoring}
\alias{consistency_scoring}
\title{Calculate consistency scores}
\usage{
consistency_scoring(pfp, pdir = NULL, method = "euclidean",
  fmt = "Luv")
}
\arguments{
\item{pdir}{directory path for directory to put plots in (optional)}

\item{method}{method for use in color space distance calculations
("euclidean" for Euclidean distances, "manhattan" for
Manhattan/city block distances - "euclidean" is standard)}

\item{fmt}{fmt: specifies color format/space to be used ("sRGB", "Luv")}

\item{dfp}{participant data csv file path (has to be formatted with
comma "," separator, and have a header), e. g.
"/Users/myname/Data_folder/bokstaverochsiffror.csv"}

\item{savepath}{(optional) file path for saving produced plot}
}
\description{
Calculates consistency scores using data from a consistency
test for synesthesia. Can also create plots for each data set using the
plot_syn_cons function.
}
\seealso{
\code{\link[synratss]{prop_col}} \code{\link[synratss]{dist_sum}}
\code{\link[synratss]{plot_syn_cons}}
}
